Add D2T3 to list of properties copied over during UpdateID
