+++ EGG9000.Bot/Commands/MiscCommandsSlash.cs
@@ -8,33 +8,16 @@ using EGG9000.Bot.Helpers;
 using EGG9000.Common.Helpers;
 using Humanizer;
 using Microsoft.EntityFrameworkCore;
-using Newtonsoft.Json;
 using System;
-using System.Collections.Generic;
-using System.IO;
 using System.Linq;
-using System.Net.Http;
-using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
-using static EGG9000.Bot.Helpers.FixedWidthTable;
 using static EGG9000.Common.Helpers.Prefarm;
 using EGG9000.Common.Services;
 using EGG9000.Common.Commands;
-using EGG9000.Common.Extensions;
-using EGG9000.Common.JsonData.EiAfxData;
 using Microsoft.Extensions.Logging;
-using Microsoft.Extensions.Primitives;
-using System.Globalization;
-using Google.Protobuf.WellKnownTypes;
-using static Ei.Backup.Types;
-using static EGG9000.Bot.Commands.ContractCommandsSlash;
-using System.ComponentModel;
+using static EGG9000.Common.Helpers.Discord.EmbedHelpers;
 using EGG9000.Bot.Automated.Coops;
-using Microsoft.AspNetCore.Connections.Features;
-using static EGG9000.Bot.Commands.DiscordEnums.AutoCompleteHandlers;
 using EGG9000.Bot.Common.Helpers;
-using System.Threading.Channels;
 using static EGG9000.Bot.Helpers.DiscordHelpersExt;
 
 namespace EGG9000.Bot.Commands {
