+++ EGG9000.Bot/Commands/StaffCommands.cs
@@ -4,44 +4,29 @@ using Discord.WebSocket;
 using EGG9000.Bot.Automated;
 using EGG9000.Common.Database;
 using EGG9000.Common.Database.Entities;
-using EGG9000.Bot.EggIncAPI;
-using EGG9000.Bot.Helpers;
-
 using EGG9000.Common.Helpers;
 
 using Humanizer;
 
 using Microsoft.EntityFrameworkCore;
 
-using Newtonsoft.Json;
-
-
 using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
-using System.Net.Http;
 using System.Text;
-using System.Text.RegularExpressions;
 using System.Threading.Tasks;
 
 using static EGG9000.Bot.Helpers.FixedWidthTable;
 using static EGG9000.Common.Helpers.Prefarm;
-using static EGG9000.Bot.Commands.ContractCommandsSlash;
+using static EGG9000.Common.Helpers.Discord.EmbedHelpers;
+using static EGG9000.Bot.Commands.DiscordEnums.AutoCompleteHandlers;
 using EGG9000.Common.Services;
-using RazorEngine.Compilation.ImpromptuInterface.Dynamic;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using System.Diagnostics;
-using System.ServiceProcess;
-using Microsoft.EntityFrameworkCore.Metadata.Internal;
 using EGG9000.Common.Commands;
-using EGG9000.Common.Contracts;
-using System.Collections;
-using System.Numerics;
 using EGG9000.Bot.Services;
-using static EGG9000.Bot.Commands.DiscordEnums.AutoCompleteHandlers;
-using System.Linq.Expressions;
 
 namespace EGG9000.Bot.Commands {
     public static class StaffCommands {
