+++ EGG9000.Bot/Services/CommandService.cs
@@ -1,11 +1,8 @@
 ﻿
 using Discord;
-using Discord.Net;
-using Discord.Rest;
 using Discord.WebSocket;
 
 using EGG9000.Bot.Automated;
-using EGG9000.Common.Helpers;
 using EGG9000.Common.Services;
 using EGG9000.Bot.Helpers;
 using EGG9000.Bot.Commands;
@@ -18,29 +15,23 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Microsoft.IdentityModel.Tokens;
-
-using Newtonsoft.Json;
 
 using System;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Reflection;
-using System.Text;
 using System.Text.RegularExpressions;
 using System.Threading;
 using System.Threading.Tasks;
 
 using static EGG9000.Common.Services.FauxCommand;
-using System.IO.Pipes;
 using System.IO;
 using MassTransit;
 using EGG9000.Bot.Consumers;
 using EGG9000.Common.Extensions;
 using EGG9000.Bot.Automated.Coops;
-using static Microsoft.EntityFrameworkCore.DbLoggerCategory.Database;
-using static EGG9000.Bot.Commands.ContractCommandsSlash;
+using static EGG9000.Common.Helpers.Discord.EmbedHelpers;
 
 namespace EGG9000.Bot.Services {
